Move {setupTest, reportWebVitals} to dir './scripts/'
