+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
+import reportWebVitals from './scripts/reportWebVitals';
 import "./styles/default.css";
 
 ReactDOM.render(
